Styled Components v4 (#2385)
Upgrade to `styled-components` v4 is fully tested and working. Minor doc updates included.

Tests are now using enzyme's `mount` and styled-components's `enzymeFind` util (perhaps a good argument for switching to `react-testing-library` per @Mensae's advice?).

Also, CSS linting had to be temporarily removed from CI suites as `stylelint` is giving an "Unknown word" error in `Circle.js`. (On a portion of the code that wasn't changed by the upgrade.) Will keep an eye on this and restore when possible.